requirements.txt
Flask==3.1.0
Flask_Cors==3.0.10
matplotlib==3.10.0
numpy==2.2.2
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.1.0
scikit_learn==1.6.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.5
